+++ payjoin/src/receive/error.rs
@@ -36,13 +36,13 @@ impl From<InternalRequestError> for Error {
 }
 
 #[cfg(feature = "v2")]
-impl From<crate::v2::HpkeError> for Error {
-    fn from(e: crate::v2::HpkeError) -> Self { Error::Server(Box::new(e)) }
+impl From<crate::hpke::HpkeError> for Error {
+    fn from(e: crate::hpke::HpkeError) -> Self { Error::Server(Box::new(e)) }
 }
 
 #[cfg(feature = "v2")]
-impl From<crate::v2::OhttpEncapsulationError> for Error {
-    fn from(e: crate::v2::OhttpEncapsulationError) -> Self { Error::Server(Box::new(e)) }
+impl From<crate::ohttp::OhttpEncapsulationError> for Error {
+    fn from(e: crate::ohttp::OhttpEncapsulationError) -> Self { Error::Server(Box::new(e)) }
 }
 
 /// Error that may occur when the request from sender is malformed.
