+++ payjoin/src/receive/v2/error.rs
@@ -1,7 +1,7 @@
 use core::fmt;
 use std::error;
 
-use crate::v2::OhttpEncapsulationError;
+use crate::ohttp::OhttpEncapsulationError;
 
 #[derive(Debug)]
 pub struct SessionError(InternalSessionError);
@@ -11,14 +11,14 @@ pub(crate) enum InternalSessionError {
     /// The session has expired
     Expired(std::time::SystemTime),
     /// OHTTP Encapsulation failed
-    OhttpEncapsulationError(OhttpEncapsulationError),
+    OhttpEncapsulation(OhttpEncapsulationError),
 }
 
 impl fmt::Display for SessionError {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
         match &self.0 {
             InternalSessionError::Expired(expiry) => write!(f, "Session expired at {:?}", expiry),
-            InternalSessionError::OhttpEncapsulationError(e) =>
+            InternalSessionError::OhttpEncapsulation(e) =>
                 write!(f, "OHTTP Encapsulation Error: {}", e),
         }
     }
@@ -28,7 +28,7 @@ impl error::Error for SessionError {
     fn source(&self) -> Option<&(dyn error::Error + 'static)> {
         match &self.0 {
             InternalSessionError::Expired(_) => None,
-            InternalSessionError::OhttpEncapsulationError(e) => Some(e),
+            InternalSessionError::OhttpEncapsulation(e) => Some(e),
         }
     }
 }
@@ -39,6 +39,6 @@ impl From<InternalSessionError> for SessionError {
 
 impl From<OhttpEncapsulationError> for SessionError {
     fn from(e: OhttpEncapsulationError) -> Self {
-        SessionError(InternalSessionError::OhttpEncapsulationError(e))
+        SessionError(InternalSessionError::OhttpEncapsulation(e))
     }
 }
